breaking out example pricers from lattice
