+++ lattice/lattice.cpp
@@ -24,6 +24,8 @@
 #include "../Vectorisation/VecX/span.h"
 
 
+#include "lattice_tools.h"
+#include "pricers.h"
 
                                                                    
 
@@ -76,465 +78,6 @@ precision of results compare for float types is incorrect.
 #include "utils.h"
 
 
-void doScan()
-{
-
-	auto v1 = std::vector<VecXX::SCALA_TYPE>(4000, 0.0);
-	FLOAT last = 0.0;// -1.0;
-	for (auto& x : v1)
-	{
-		x = last + 1.0;
-		last = x;
-	}
-	// getRandomShuffledVector(200);
-	VecXX a(v1);
-	std::vector<FLOAT> dbg = v1;
-
-	auto sum = [](auto X, auto Y) { return  X + Y; };
-
-	double time = 0;
-	{
-		TimerGuard timer(time);
-
-		VecXX result;
-		for (long l = 0; l < 1000000; ++l)
-		{
-			result = ApplyScan(a, sum);
-			//dbg = a;
-		}
-
-		dbg = result;
-	}
-	std::cout << "run time = " << time << "\n";
-
-
-
-	///////TRANSFORM SCAN //////
-	
-	//{
-	//	TimerGuard timer(time);
-	//	auto doubleIt= [](auto X) { return  X*2.; };
-	//	VecXX result;
-	//	for (long l = 0; l < 1000000; ++l)
-	//	{
-	//		result = ApplyTransformScan(a, sum, doubleIt);
-	//		//dbg = a;
-	//	}
-	//	dbg = result;
-	//}
-
-	//////////
-
-
-	
-	auto dbg_cpy = dbg;
-	time = 0;
-	{
-		TimerGuard timer(time);
-		
-		for (long l = 0; l < 1000000; ++l)
-		{
-			std::inclusive_scan(begin(v1), end(v1), dbg.begin());
-		}
-	}
-	std::cout << "run time std::scan = " << time << "\n";
-
-	for (size_t i = 0; i < dbg_cpy.size(); ++i)
-	{
-		std::cout << i << "  ,  " << dbg[i]  << "   , " << dbg_cpy[i]  << std::endl;
-	}
-
-}
-
-void testSampler()
-{
-
-	FLOAT sampleData[] = { 0., 1.0,2.,3.,4.,5.,6.,7.,8.,9.,10. };
-
-	TrinomialSampler<VecXX::INS>  zzz;
-
-	zzz.load(&sampleData[1]);
-
-	auto x = zzz.X_Minus_1.value;
-	auto y = zzz.X_0.value;
-	auto z = zzz.X_1.value;
-
-	ignore(x);
-	ignore(y);
-	ignore(z);
-}
-
-
-
-
-void doStridedSpan()
-{
-
-	auto v1 = std::vector<VecXX::SCALA_TYPE>(4000, 0.0);
-	FLOAT last = 0.0;// -1.0;
-	for (auto& x : v1)
-	{
-		x = last + 1.0;
-		last = x;
-	}
-	// getRandomShuffledVector(200);
-	VecXX a_vec(v1);
-	std::vector<FLOAT> dbg = v1;
-
-	VecXX result = a_vec;
-
-	result *= 0.0;
-
-	StridedSampler<VecXX::INS> strided_sampler(2);
-
-	//sampler.load(a_vec.data());
-
-
-
-	//DRC::Span< VecXX::INS>
-	SpanXX spn(a_vec.start(), 800);
-
-	SpanXX spnPlus(result.start() , 800);
-
-	//UnitarySampler<VecXX::INS> identity_sampler;
-
-
-	auto SQR = [](StridedSampler<VecXX::INS>& sampler)
-	{
-		auto x = sampler.X_0.value;//get<0>();
-		return x * x;
-	};
-
-
-	//need a ptr diff  betwee start span and start vecXX
-	//spnPlus *= 0;
-
-	ApplyTransformUR_X_Impl_EX(spn, spnPlus, SQR, strided_sampler, 0, int(spn.paddedSize()));// int impSZ = -1)`
-
-	std::vector<double> ddbbgg = spn;
-	ddbbgg = spnPlus;
-
-	VecXX rootvec = result;
-
-	//auto SQRT = []( VecXX::INS x) { return sqrt(x); };
-
-	//auto SQRT = [](UnitarySampler<VecXX::INS>& sampler)
-	//{
-	//	auto x = sampler.get<0>();
-	//	return sqrt(x);
-	//};
-	
-
-	auto SQRT2 = [](auto& x)
-	{
-		return sqrt(x);
-	};
-
-	auto rooted = spnPlus;
-	transform(SQRT2, spnPlus,rooted);// rootvec);
-
-	ddbbgg = rooted;
-
-
-
-	//reduce
-	//auto SUM = [](auto x, auto y) {return x + y; };
-
-	//auto val =ApplyAccumulate2UR_X(spn, SUM);
-	// goes to reduce(span,lambda)
-
-	//auto SUM_SPARSE = [](StridedSampler<VecXX::INS>& sampler_lhs, StridedSampler<VecXX::INS>& sampler_rhs)
-	//{
-	//	auto x = sampler_lhs.get<0>();
-	//	auto y = sampler_rhs.get<0>();
-	//	return x + y;
-	//};
-
-	//auto val = reduce(spn, SUM_SPARSE);
-
-	/*
-
-	//TO DO transformReduce 
-
-	auto SQR_VAL = [](auto x) {return x * x; };
-
-	const auto& cnstSpn = spn;
-
-	//auto sumSqrsXXX = ApplyTransformAccumulate2UR_X_Impl(cnstSpn, SQR_VAL, SUM);
-	//auto sumSqrsXXX = ApplyTransformAccumulate2UR_X_Impl(spn, SQR_VAL, SUM);
-
-
-	auto sumSqrsXXX = transformReduce(spn, SQR_VAL, SUM);
-
-	*/
-
-}
-
-
-
-void doTransformWithASpan()
-{
-
-	auto v1 = std::vector<VecXX::SCALA_TYPE>(4000, 0.0);
-	FLOAT last = 0.0;// -1.0;
-	for (auto& x : v1)
-	{
-		x = last + 1.0;
-		last = x;
-	}
-	// getRandomShuffledVector(200);
-	VecXX a_vec(v1);
-	std::vector<FLOAT> dbg = v1;
-
-	VecXX result = a_vec;
-
-
-	//DRC::Span< VecXX::INS>
-	 SpanXX spn(a_vec.start(), 200);
-
-	 SpanXX spnPlus(a_vec.start()+200, 200);
-
-	UnitarySampler<VecXX::INS> identity_sampler;
-
-	
-	auto SQR = [](UnitarySampler<VecXX::INS>& sampler)
-	{
-		auto x = sampler.X_0.value; // get<0>();
-		return x * x; 
-	};
-
-
-	auto SQR2 = [](auto x)
-	{
-		return x * x;
-	};
-	
-
-
-	//need a ptr diff  betwee start span and start vecXX
-
-	ApplyTransformUR_X_Impl_EX(a_vec, spnPlus, SQR, identity_sampler, 0, int(spn.paddedSize()));// int impSZ = -1)`
-
-	ApplyTransformUR_X_Impl_EX(spn, spnPlus, SQR, identity_sampler, 0, int(spn.paddedSize()));// int impSZ = -1)
-
-	//trapped into using same container , ie vecxx vecxx  , not mixed so add extra type argument 
-	std::vector<double> ddbbgg = spn;
-
-	transform(SQR2, spn, spnPlus);
-
-	ddbbgg  = spnPlus;
-
-
-	auto isEven = [](auto x) { return !((x - VecXX::INS((2.0) * floor(x * VecXX::INS(0.5)))) > VecXX::INS(0.)); };
-	auto evenView = filter(isEven, a_vec);
-
-	ddbbgg = evenView;
-
-	ApplyTransformUR_X_Impl_EX(spn, evenView, SQR, identity_sampler, 100, int(spn.paddedSize()));
-
-	ddbbgg = evenView;
-
-
-	VecXX anotherVector(v1);
-	VecXX resvec(v1);
-
-
-	SpanXX smallSpan(anotherVector.start() + 20, 100);
-
-	auto vwOfSpan = filter(isEven, smallSpan);
-
-	ddbbgg = vwOfSpan;
-
-	auto CUBE_IT = [](auto x) { return x * x * x; };
-
-	auto neVw = vwOfSpan;
-	transformM(CUBE_IT, neVw);// , resvec);// vwOfSpan); // broken somehow
-
-	for (auto& x : neVw)
-	{
-		std::cout << x << std::endl;
-	}
-
-	ddbbgg = neVw;
-
-	ddbbgg = vwOfSpan;
-
-	//transforming view as a result of filtering a span
-
-	vwOfSpan =transform(CUBE_IT, vwOfSpan); // broken somehow
-
-	ddbbgg = vwOfSpan;
-
-	for (auto& x : vwOfSpan)
-	{
-		std::cout << x << std::endl;
-	}
-
-
-
-	//reduce
-	//auto SUM = [](auto x, auto y) {return x + y; };
-
-	//auto val =ApplyAccumulate2UR_X(spn, SUM);
-	// goes to reduce(span,lambda)
-
-
-	// possible problem calls align load
-	// we need a more general version using unaligne dor sampled load
-	//auto val = reduce(spn, SUM);
-
-
-
-	//TO DO transformReduce 
-
-//	auto SQR_VAL = [](auto x) {return x * x; };
-
-//	const auto& cnstSpn = spn;
-
-	//auto sumSqrsXXX = ApplyTransformAccumulate2UR_X_Impl(cnstSpn, SQR_VAL, SUM);
-	//auto sumSqrsXXX = ApplyTransformAccumulate2UR_X_Impl(spn, SQR_VAL, SUM);
-
-
-//	auto sumSqrsXXX = transformReduce(spn, SQR_VAL, SUM);
-
-
-
-	//auto trform = ApplyUnitaryOperation(cnstSpn, SQR_VAL);
-	//return trform.getScalarValue();
-
-
-	//need to make apply unitary op work with generic vec/view so that it works with span
-
-
-
-
-}
-
-
-void doZipping()
-{
-
-	auto v1 = getRandomShuffledVector(200);
-	VecXX a(v1);
-	VecXX b = 2. * a;
-	VecXX c = 3. * a;
-	VecXX d = 4. * a;
-	VecXX e = 5. * a;
-
-	auto zipped = make_Zipped<VecXX::INS>(a, b, c);
-	auto it = make_Zipped_itr<VecXX::INS>(a, b, c);
-
-
-	//auto zipped =
-	const VecXX& aa = a;
-	const VecXX& bb = b;
-	const VecXX& cc = c;
-	const VecXX& dd = d;
-	const VecXX& ee = e;
-
-	auto zp5 = make_Zipped<VecXX::INS>(aa, bb, cc, dd, ee);
-	auto zp5_itr = make_Zipped_itr<VecXX::INS>(aa, bb, cc, dd, ee);
-
-	ignore(zp5);
-	ignore(zp5_itr);
-
-
-	//it.inc(1);
-
-	auto addingLambda = [](auto& zpped)
-	{
-		const auto& X = std::get<0>(zpped.m_registers);
-		const auto& Y = std::get<1>(zpped.m_registers);
-		const auto& Z = std::get<2>(zpped.m_registers);
-
-		return X + Y + Z;
-	};
-
-
-	VecXX::INS X = 0.;
-	for (int i = 0; i < 200 / 8; ++i)
-	{
-		zipped.load(it);
-		auto sum = addingLambda(zipped);
-		X += sum;
-		it.inc(1);
-	}
-
-	std::cout << X[0];
-
-	zipped.load(it);
-
-
-	auto addingLambda5 = [](auto& zpped)
-	{
-		const auto& X = std::get<0>(zpped.m_registers);
-		const auto& Y = std::get<1>(zpped.m_registers);
-		const auto& Z = std::get<2>(zpped.m_registers);
-		const auto& L = std::get<3>(zpped.m_registers);
-		const auto& M = std::get<4>(zpped.m_registers);
-
-		return X + Y + Z + L + M;
-	};
-
-
-
-	//Vec<INS_VEC> res =   transform(OP & oper, const Zipped_ITR< INS_VEC, N>&zip)
-
-	auto res = transform(addingLambda5, zp5_itr);
-
-	std::vector<FLOAT> vdb = res;
-
-	enum class Access { down = -1, mid = 0, up = 1 };
-	//	using Mysample = Named_Zip_iter < VecXX::INS, 3, Access > ;
-
-		//NAMED_INDEX
-
-		//Mysample instance;
-
-
-	//	Named_Zipped_Reg < VecXX::INS, 3, Access > myReg;
-	Zipped_Reg < VecXX::INS, 3 > myReg;
-	//	using reg = Zipped_Reg < VecXX::INS, 3 >;
-
-		//auto zz = Access::down;
-	auto yy = myReg.get((int)Access::up);
-	ignore(yy);
-
-
-
-	//////////////////////////////
-
-
-	VecXX out1 = a;
-	VecXX out2 = b;
-
-	VecXX& out1_r = out1;
-	VecXX& out2_r = out2;
-
-
-	auto zp_out = make_Zipped_itr_ref<VecXX::INS>(out1_r, out2_r);
-
-	auto addingLambdaIO = [&](const auto& zpped, auto& out_zipp)
-	{
-		const auto& X = std::get<0>(zpped.m_registers);
-		const auto& Y = std::get<1>(zpped.m_registers);
-		const auto& Z = std::get<2>(zpped.m_registers);
-
-		std::get<0>(out_zipp.m_registers) = X + Y + Z;
-		std::get<1>(out_zipp.m_registers) = X + Y;
-	};
-
-
-	transform(addingLambdaIO, zp5_itr, zp_out);
-
-
-	std::vector<FLOAT> vdb1 = out1;
-	std::vector<FLOAT> vdb2 = out2;
-
-}
-
-
-
 VecXX  blackScholes(const VecXX& S, const VecXX& K, const VecXX& t, const VecXX& r, const VecXX& sigma)
 {
 	auto invK = 1.0 / K;
@@ -555,59 +98,8 @@ VecXX  blackScholes(const VecXX& S, const VecXX& K, const VecXX& t, const VecXX&
 }
 
 
-double europeanBinomialPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	VecXX terminalAssetPrices(1.0, N + 1);
-
-	double Dt = T / N;
-	double  u = std::exp(sig * std::sqrt(Dt));
-	double d = 1. / u;
-
-
-	VecXX::INS pu = (exp(r * Dt) - d) / (u - d);
-	VecXX::INS oneMinusP = (1.0 - pu);
-	VecXX::INS disc = exp(-r * Dt);
-
-	BinomialSampler<VecXX::INS> sampler;
-
-	auto binomialRollBack = [=](BinomialSampler<VecXX::INS>& sampler)
-	{
-		const auto& X1 = sampler.X_1.value;
-		const auto& X0 = sampler.X_0.value;
-		return disc * (X1 * pu + X0 * oneMinusP);
-	};
-
-
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-	//	auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * std::pow(d, N + 2);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= (u * u);
-		el = last;
-	}
-
-
-	auto odd_slice = transform(payOffFunc, terminalAssetPrices);
-	auto even_slice = odd_slice;
-
-	int j = N + 1;
-	for (int i = 0; i < N / 2; ++i)
-	{
-		transform(odd_slice, even_slice, binomialRollBack, sampler, 0, j);
-		transform(even_slice, odd_slice, binomialRollBack, sampler, 0, j - 1);
-		j -= 2;
-
-	}
-	return odd_slice[0];;
-}
-
 void doBinomialPricer()
 {
-
 
 	double res = 0.0;
 	double time = 0;
@@ -620,8 +112,6 @@ void doBinomialPricer()
 	{
 		TimerGuard timer(time);
 
-
-
 		int N = 30000;
 
 		res = europeanBinomialPricer(S, K, vol, rate, T, N);
@@ -632,219 +122,6 @@ void doBinomialPricer()
 	std::cout << " takes secs" << time << "\n";
 }
 
-double europeanTrinomialPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	double y = 0.0;// 0.03; //div yield
-
-
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(3.0* Dt);
-	
-	double v = r - y - 0.5 * sig * sig;
-
-
-
-	VecXX::INS pu = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) + (v * Dt) / Dx);
-	VecXX::INS pd = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) - (v * Dt) / Dx);
-	VecXX::INS pm = 1. - (Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx);
-
-	VecXX::INS disc = exp(-r * Dt);
-
-	TrinomialSampler<VecXX::INS> sampler;
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-
-		auto X1 = sampler.X_Minus_1.value;
-		auto X0 = sampler.X_0.value;
-		auto X_1 = sampler.X_1.value;
-
-		return disc * (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	auto odd_slice = transform(payOffFunc, terminalAssetPrices);
-	auto even_slice = odd_slice;
-
-	int j = 2 * N + 1 - 1;
-	int i = 0;
-	for (; i < N; i += 2)
-	{
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, i, j);
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-		j -= 2;
-	}
-
-	return odd_slice[N];
-}
-
-double americanTrinomialPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	double y = 0.0;// 0.03; //div yield
-
-
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-
-	VecXX::INS pu = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) + (v * Dt) / Dx);
-	VecXX::INS pd = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) - (v * Dt) / Dx);
-	VecXX::INS pm = 1. - (Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx);
-
-
-	VecXX::INS disc = exp(-r * Dt);
-	TrinomialSampler<VecXX::INS> sampler;
-
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-		auto X1 = sampler.X_Minus_1.value;
-		auto X0 = sampler.X_0.value;
-		auto X_1 = sampler.X_1.value;
-		return disc * (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K -X , 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-	auto odd_slice = excerciseValue;
-
-	UnitarySampler<VecXX::INS> identity_sampler; //identity just  passes through
-
-	auto applyEarlyExcercise = [=](UnitarySampler<VecXX::INS>& sampler, auto excercisePrice)
-	{
-		auto optPrice = sampler.X_0.value; //.get<0>();
-		return max(optPrice, excercisePrice);
-	};
-
-
-	auto even_slice = odd_slice;
-
-	int j = 2 * N + 1 - 1;
-	int i = 0;
-	for (; i < N; i += 2)
-	{
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, i, j);
-		// transform to get early excercise for american bit , iderntity sampler just passes values straight through
-		transform(even_slice, excerciseValue, even_slice, applyEarlyExcercise, identity_sampler, i, j);
-
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-		transform(odd_slice, excerciseValue, odd_slice, applyEarlyExcercise, identity_sampler, i + 1, j - 1);
-
-		j -= 2;
-	}
-
-	return odd_slice[N];
-}
-
-double europeanTrinomialPricer1(double S, double K, double sig, double r, double T, int N)
-{
-
-	double y = 0.0;// 0.03; //div yield
-
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-	//double  u = Dx;
-	//double d = 1. / u;
-
-
-	VecXX::INS pu = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) + (v * Dt) / Dx);
-	VecXX::INS pd = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) - (v * Dt) / Dx);
-	VecXX::INS pm = 1. - (Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx);
-
-
-	VecXX::INS disc = exp(-r * Dt);
-	TrinomialSampler<VecXX::INS> sampler;
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-		const auto& X1 = sampler.X_1.value;
-		const auto& X0 = sampler.X_0.value;
-		const auto& X_1 = sampler.X_Minus_1.value;
-		return disc * (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K -X , 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-	auto odd_slice = excerciseValue;
-
-	UnitarySampler<VecXX::INS> identity_sampler; //identity just 
-
-//	auto applyEarlyExcercise = [=](UnitarySampler<VecXX::INS>& sampler, auto excercisePrice)
-//	{
-//		auto optPrice = sampler.get<0>();
-//		return max(optPrice, excercisePrice);
-//	};
-
-
-	auto even_slice = odd_slice;
-
-	int j = 2 * N + 1 - 1;
-	int i = 0;
-	for (; i < N; i += 2)
-	{
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, i, j);
-		// transform to get early excercise for american bit , iderntity sampler just passes values straight through
-		//transform(even_slice, excerciseValue, even_slice, applyEarlyExcercise, identity_sampler, i, j);
-
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-    	//transform(odd_slice, excerciseValue, odd_slice, applyEarlyExcercise, identity_sampler, i + 1, j - 1);
-
-		j -= 2;
-	}
-
-	return odd_slice[N];
-}
 
 void doTrinomialPricer()
 {
@@ -869,92 +146,6 @@ void doTrinomialPricer()
 	std::cout << "trinomial takes secs" << time << "\n";
 }
 
-double americanFiniteDiffPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	//dividend yield
-	double y = 0.0;
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-
-	VecXX::INS pu = 0.5 * Dt * ((sig * sig) / (Dx * Dx) + v / Dx);
-	VecXX::INS pd = 0.5 * Dt * ((sig * sig) / (Dx * Dx) - v / Dx);
-	VecXX::INS pm = 1. - Dt * (sig * sig) / (Dx * Dx) - r * Dt;
-
-
-
-	TrinomialSampler<VecXX::INS> sampler;
-	//introduces offset variables so that we can get vectorised versions
-	// of X[i+1], X[i] and  x[i-1]
-	// under the hood these do unaligned loads into registers so taht we can still
-	// apply the vectorised 
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-		const auto& X1 = sampler.X_1.value;
-		const auto& X0 = sampler.X_0.value;
-		const auto& X_1 = sampler.X_Minus_1.value;
-		return  (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-	//Pay off functions
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-	auto odd_slice = excerciseValue;
-
-	UnitarySampler<VecXX::INS> identity_sampler; //identity  pas through
-
-	//this is the american part of the option excercise
-	auto applyEarlyExcercise = [=](UnitarySampler<VecXX::INS>& sampler, auto excercisePrice)
-	{
-		auto optPrice = sampler.X_0.value; //.get<0>();
-		return max(optPrice, excercisePrice);
-	};
-
-	auto even_slice = odd_slice * 0.0;
-
-	int J = 2 * N;
-	int k = 0;
-	for (; k < N; k += 2)
-	{
-
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, 0, J);
-
-		//apply boundary condition
-		even_slice[0] = even_slice[1] + terminalAssetPrices[1] - terminalAssetPrices[0];
-		even_slice[J] = even_slice[J - 1];
-		// transform to get early excercise for american exercise , iderntity sampler just passes values straight through
-		transform(even_slice, excerciseValue, even_slice, applyEarlyExcercise, identity_sampler, 0, J);
-
-
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, 0, J);
-		//boundary condition
-		odd_slice[0] = odd_slice[1] + terminalAssetPrices[1] - terminalAssetPrices[0];
-		odd_slice[J] = odd_slice[J - 1];
-		transform(odd_slice, excerciseValue, odd_slice, applyEarlyExcercise, identity_sampler, 0, J);
-
-	}
-
-	return odd_slice[N];
-}
 
 void doAmericanFiniteDiff()
 {
@@ -978,273 +169,6 @@ void doAmericanFiniteDiff()
 	std::cout << "finite dif takes secs" << time << "\n";
 }
 
-double americanImplicitFiniteDiffPricerFast(double S, double K, double sig, double r, double T, int N)
-{
-
-	
-	double y = 0.0;//dividend yield
-
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-
-	VecXX::SCALA_TYPE pu = -0.5 * Dt * ((sig * sig) / (Dx * Dx) + v / Dx);
-	VecXX::SCALA_TYPE pd = -0.5 * Dt * ((sig * sig) / (Dx * Dx) - v / Dx);
-	VecXX::SCALA_TYPE pm = 1. + Dt * (sig * sig) / (Dx * Dx) + r * Dt;
-
-
-	std::vector<FLOAT> vdbg;
-	//Pay off functions
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	//option value at maturity
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-
-	auto american = [](auto X, auto Y) { return select(X > Y, X, Y); };
-
-
-	//derivative boundary condition
-	double  lambda_L = -1. * (terminalAssetPrices[1] - terminalAssetPrices[0]);
-	double  lambda_U = 0.0;
-
-	auto odd_slice = excerciseValue;
-	//	vdbg = odd_slice;
-
-
-	auto even_slice = odd_slice * 0.0;
-
-	int J = 2 * N;
-	int k = 0;
-
-	VecXX pmp(1.0, J + 1);
-	VecXX pp(1.0, J + 1);
-
-	////////////
-	//LOOP HOIST BITS FROM  IMPLICIT TRIDIAGONAL 
-
-	pmp[1] = pm + pd;
-	pp[1] = odd_slice[1] + pd * lambda_L;
-
-
-	auto pu_pd = pu * pd;
-
-	// eliminate upper diagonal
-	for (int j = 2; j < J; ++j)
-	{
-		pmp[j] = pm - pu_pd / pmp[j - 1];
-	}
-
-	auto inv_pmp = 1.0 / pmp;
-
-	auto pd_inv_pmp = pd * inv_pmp;
-
-	/////////////
-
-	for (; k < N; k += 2)
-	{
-
-		// SOLVE IMPLICIT TRIDIAGONAL  IN LINE 	SUB BOUNDARY CONDITION AT J = -n INTO  J = -n+1
-	
-		//pmp[1] = pm + pd;
-		pp[1] = odd_slice[1] + pd * lambda_L;
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-
-			pp[j] = odd_slice[j] - pp[j - 1] * pd_inv_pmp[j - 1];
-		}
-
-		even_slice[1] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		even_slice[J - 1] = even_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 2; j != 0; j--)
-		{
-			even_slice[j] = (pp[j] - pu * even_slice[j + 1]) * inv_pmp[j];
-		}
-
-		//american excercise bit
-		even_slice = transform(american, even_slice, (const VecXX&)excerciseValue);
-
-
-		// now calculate the  odd slice 
-
-		pp[1] = even_slice[1] + pd * lambda_L;
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-			pp[j] = even_slice[j] - pp[j - 1] * pd_inv_pmp[j - 1];
-		}
-
-		odd_slice[1] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		odd_slice[J - 1] = odd_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 2; j != 0; j--)
-		{
-			odd_slice[j] = (pp[j] - pu * odd_slice[j + 1]) * inv_pmp[j];
-		}
-
-		//american excercise bit
-		odd_slice = transform(american, odd_slice, (const VecXX&)excerciseValue);
-	}
-
-	return odd_slice[N];
-}
-
-double americanImplicitFiniteDiffPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	//dividend yield
-	double y = 0.0;// 0.03;// 0.03; //div yield
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-	//double  u = Dx;
-	//double d = 1. / u;
-
-
-	VecXX::SCALA_TYPE pu = -0.5 * Dt * ((sig * sig) / (Dx * Dx) + v / Dx);
-	VecXX::SCALA_TYPE pd = -0.5 * Dt * ((sig * sig) / (Dx * Dx) - v / Dx);
-	VecXX::SCALA_TYPE pm = 1. + Dt * (sig * sig) / (Dx * Dx) + r * Dt;
-
-
-	std::vector<FLOAT> vdbg;
-	//Pay off functions
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	//option value at maturity
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-
-	//derivative boundary condition
-	double  lambda_L = -1. * (terminalAssetPrices[1] - terminalAssetPrices[0]);
-	double  lambda_U = 0.0;
-
-	auto odd_slice = excerciseValue;
-	vdbg = odd_slice;
-
-
-	auto even_slice = odd_slice * 0.0;
-
-	int J = 2 * N;
-	int k = 0;
-
-	VecXX pmp(1.0, J + 1);
-	VecXX pp(1.0, J + 1);
-
-	for (; k < N; k += 2)
-	{
-
-		// SOLVE IMPLICIT TRIDIAGONAL  IN LINE 
-		//SUB BOUNDARY CONDITION AT J = -n INTO  J = -n+1
-		pmp[1] = pm + pd;
-		pp[1] = odd_slice[1] + pd * lambda_L;
-
-
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-			pmp[j] = pm - pu * pd / pmp[j - 1];
-			pp[j] = odd_slice[j] - pp[j - 1] * pd / pmp[j - 1];
-		}
-
-		even_slice[1] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		even_slice[J - 1] = even_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 2; j != 0; j--)
-		{
-			even_slice[j] = (pp[j] - pu * even_slice[j + 1]) / pmp[j];
-		}
-
-		/*
-
-		// american
-		for (int j = 0; j < (J+1); j++)
-		{
-			even_slice[j] = std::max(even_slice[j], excerciseValue[j]);
-		}
-
-*/
-//even_slice =	transform(even_slice, excerciseValue);
-
-// calc odd slice now
-//////////////////////////////////////////////////////////////////////
-
-		pmp[1] = pm + pd;
-		pp[1] = even_slice[1] + pd * lambda_L;
-
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-			pmp[j] = pm - pu * pd / pmp[j - 1];
-			pp[j] = even_slice[j] - pp[j - 1] * pd / pmp[j - 1];
-		}
-
-		odd_slice[1] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		odd_slice[J - 1] = odd_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 2; j != 0; j--)
-		{
-			odd_slice[j] = (pp[j] - pu * odd_slice[j + 1]) / pmp[j];
-		}
-
-		/*
-		//american
-		for (int j = 0; j < (J + 1); j++)
-		{
-			odd_slice[j] = std::max(odd_slice[j], excerciseValue[j]);
-		}
-		*/
-
-	}
-
-	return odd_slice[N];
-}
 
 void doAmericanImplicitFiniteDiff()
 {
@@ -1268,151 +192,6 @@ void doAmericanImplicitFiniteDiff()
 }
 
 
-/////////////////////
-
-//still broken ???
-double americanCrankNicholsonPricer(double S, double K, double sig, double r, double T, int N)
-{
-
-	//dividend yield
-	double y = 0.;// 0.03;// 0.0;// 0.03;// 0.03; //div yield
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(1.0 * Dt);// 0.2;//
-	double v = r - y - 0.5 * sig * sig;
-
-
-	VecXX::SCALA_TYPE pu = -0.25 * Dt * ((sig * sig) / (Dx * Dx) + v / Dx);
-	VecXX::SCALA_TYPE pd = -0.25 * Dt * ((sig * sig) / (Dx * Dx) - v / Dx);
-	VecXX::SCALA_TYPE pm = 1. + 0.5 * Dt * (sig * sig) / (Dx * Dx) + 0.5 * r * Dt;
-
-	std::vector<FLOAT> vdbg;
-	//Pay off functions
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	//option vakue at maturity
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-
-
-	//derivative boundary condition
-	double  lambda_L = -1. * (terminalAssetPrices[1] - terminalAssetPrices[0]);
-	double  lambda_U = 0.0;
-
-	auto odd_slice = excerciseValue;
-
-	//set up slices
-	auto even_slice = odd_slice * 0.0;
-	even_slice[0] = odd_slice[0];
-
-	int J = 2 * N;
-	int k = 0;
-
-	VecXX pmp(1.0, J + 1);
-	VecXX pp(1.0, J + 1);
-
-	for (; k <= N; k += 2)
-	{
-
-		// SOLVE IMPLICIT TRIDIAGONAL  IN LINE //SUB BOUNDARY CONDITION AT J = -n INTO  J = -n+1
-		
-		pmp[1] = pm + pd;
-		pp[1] = -pu * odd_slice[2] - (pm - 2.) * odd_slice[1] - pd * odd_slice[0] + pd * lambda_L;
-
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-			pmp[j] = pm - pu * pd / pmp[j - 1];
-			pp[j] = -pu * odd_slice[j + 1] - (pm - 2.0) * odd_slice[j] - pd * odd_slice[j - 1] - pp[j - 1] * pd / pmp[j - 1];
-		}
-
-		even_slice[J] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		even_slice[J - 1] = even_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 1; j >= 0; j--)
-		{
-			even_slice[j] = (pp[j] - pu * even_slice[j + 1]) / pmp[j];
-		}
-
-
-		even_slice[0] = odd_slice[0];
-		//vdbg = even_slice;
-/*
-		//american condition
-		for (int j = 0; j < (J+1); j++)
-		{
-			even_slice[j] = std::max(even_slice[j], excerciseValue[j]);
-		}
-*/
-
-//	vdbg = even_slice;
-
-//	vdbg = odd_slice;
-//	vdbg = pmp;
-//	vdbg = pp;
-
-
-	// calc odd slice now
-//////////////////////////////////////////////////////////////////////
-
-		pmp[1] = pm + pd;
-		pp[1] = -pu * even_slice[2] - (pm - 2.) * even_slice[1] - pd * even_slice[0] + pd * lambda_L;
-
-
-		// eliminate upper diagonal
-		for (int j = 2; j < J; ++j)
-		{
-			pmp[j] = pm - pu * pd / pmp[j - 1];
-			pp[j] = -pu * even_slice[j + 1] - (pm - 2.0) * even_slice[j] - pd * even_slice[j - 1] - pp[j - 1] * pd / pmp[j - 1];
-		}
-
-		odd_slice[J] = (pp[J - 1] + pmp[J - 1] * lambda_U) / (pu + pmp[J - 1]);
-		odd_slice[J - 1] = odd_slice[J] - lambda_U;
-
-
-		// back substitution
-		for (int j = J - 1; j >= 0; j--)
-		{
-			odd_slice[j] = (pp[j] - pu * odd_slice[j + 1]) / pmp[j];
-		}
-
-		odd_slice[0] = even_slice[0];
-		/*
-				//american condition
-				for (int j = 0; j < (J + 1); j++)
-				{
-					odd_slice[j] = std::max(odd_slice[j], excerciseValue[j]);
-				}
-			*/
-
-			//	vdbg = odd_slice;
-			//	vdbg = even_slice;
-			//	vdbg = pmp;
-			//	vdbg = pp;
-
-	}
-
-	return odd_slice[N];
-}
-
 void doAmericanCrankNicholson()
 {
 	double time = 0;
@@ -1432,99 +211,6 @@ void doAmericanCrankNicholson()
 	std::cout << std::setprecision(12) << "price " << res << "\n";
 	std::cout << " takes secs" << time << "\n";
 }
-
-double americanTrinomialPricerUpAndOut(double S, double K, double sig, double r, double T, double H, double rebate, int N)
-{
-
-	double y = 0.0;// 0.03; //div yield
-
-
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-	//double  u = Dx;
-	//double d = 1. / u;
-
-
-	VecXX::INS pu = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) + (v * Dt) / Dx);
-	VecXX::INS pd = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) - (v * Dt) / Dx);
-	VecXX::INS pm = 1. - (Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx);
-
-
-	VecXX::INS disc = exp(-r * Dt);
-	TrinomialSampler<VecXX::INS> sampler;
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-
-		auto X1 = sampler.X_Minus_1.value;
-		auto X0 = sampler.X_0.value;
-		auto X_1 = sampler.X_1.value;
-
-		return disc * (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-
-	//auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); }; //call
-	auto payOffFunc = [=](auto X) { return select(X < K, K - X, 0.0); };  //put
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-	auto odd_slice = excerciseValue;
-
-	UnitarySampler<VecXX::INS> identity_sampler; //identity just 
-
-	auto applyEarlyExcercise = [=](UnitarySampler<VecXX::INS>& sampler, auto excercisePrice)
-	{
-		auto optPrice = sampler.X_0.value; //.get<0>();
-		return max(optPrice, excercisePrice);
-	};
-
-
-	auto applyBarrier = [=](UnitarySampler<VecXX::INS>& sampler, auto stockPrice)
-	{
-		auto optPrice = sampler.X_0.value;//.get<0>();
-		return select(stockPrice < H, optPrice, rebate);
-	};
-
-
-	auto even_slice = odd_slice;
-
-	int j = 2 * N + 1 - 1;
-	int i = 0;
-	for (; i < N; i += 2)
-	{
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, i, j);
-		// transform to get early excercise for american bit , iderntity sampler just passes values straight through
-	//	transform(even_slice, excerciseValue, even_slice, applyEarlyExcercise, identity_sampler, i, j);
-		transform(even_slice, terminalAssetPrices, even_slice, applyBarrier, identity_sampler, i, j);
-
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-		//	transform(odd_slice, excerciseValue, odd_slice, applyEarlyExcercise, identity_sampler, i + 1, j - 1);
-		transform(odd_slice, terminalAssetPrices, odd_slice, applyBarrier, identity_sampler, i + 1, j - 1);
-
-		j -= 2;
-	}
-
-    ignore(applyEarlyExcercise);
-
-	return odd_slice[N];
-
-}
-
 
 void doAmericanTrinomialPricerUpAndOut()
 {
@@ -1548,122 +234,6 @@ void doAmericanTrinomialPricerUpAndOut()
 	}
 	std::cout << std::setprecision(12) << "price " << res << "\n";
 	std::cout << " takes secs" << time << "\n";
-}
-
-
-double euroTrinomialPricerWithInit(double S, double K, double sig, double r, double T, int N)
-{
-
-	double y = 0.0;// 0.03; //div yield
-	VecXX terminalAssetPrices(1.0, 2 * N + 1);
-
-	double Dt = T / N;
-	double Dx = sig * std::sqrt(2.0 * Dt);
-	double v = r - y - 0.5 * sig * sig;
-
-	VecXX::INS pu = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) + (v * Dt) / Dx);
-	VecXX::INS pd = 0.5 * ((Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx) - (v * Dt) / Dx);
-	VecXX::INS pm = 1. - (Dt * sig * sig + v * v * Dt * Dt) / (Dx * Dx);
-
-
-	VecXX::INS disc = exp(-r * Dt);
-	TrinomialSampler<VecXX::INS> sampler;
-
-	auto trinomialRollBack = [=](TrinomialSampler<VecXX::INS>& sampler)
-	{
-		const auto& X1 = sampler.X_1.value;
-		const auto& X0 = sampler.X_0.value;
-		const auto& X_1 = sampler.X_Minus_1.value;
-		return disc * (X1 * pu + X0 * pm + X_1 * pd);
-	};
-
-	//call
-	auto payOffFunc = [=](auto X) { return select(X > K, X - K, 0.0); };
-
-	//put
-	//auto payOffFunc = [=](auto X) { return select(X < K, K -X , 0.0); };
-
-	//set up underlying asset prices at maturity
-	double last = S * exp(-(N + 1) * Dx);
-	double edx = exp(Dx);
-	for (auto& el : terminalAssetPrices)
-	{
-		last *= edx;
-		el = last;
-	}
-
-	auto excerciseValue = transform(payOffFunc, terminalAssetPrices);
-	auto odd_slice = excerciseValue;
-
-
-	UnitarySampler<VecXX::INS> identity_sampler; //identity just 
-
-	auto applyEarlyExcercise = [=](UnitarySampler<VecXX::INS>& sampler, auto excercisePrice)
-	{
-		auto optPrice = sampler.X_0.value; //.get<0>();
-		return max(optPrice, excercisePrice);
-	};
-
-
-	auto even_slice = odd_slice;
-
-
-	/// blacks initialisation 
-	VecXX::INS invK = 1.0 / K;
-	VecXX::INS discountedRate = exp(-r * Dt);
-
-	VecXX::INS rootT = sqrt(Dt);
-	VecXX::INS sigmaRootT = rootT * sig;
-	VecXX::INS invSigmaRootT = 1.0 / sigmaRootT;
-	VecXX::INS halfSigmaSqrd_t = (0.5 * sig * sig + r) * Dt;
-
-	VecXX::INS Strike = K;
-
-	auto blackScholeInit = [&](VecXX::INS S)
-	{
-		VecXX::INS S_invK = S * invK;
-		VecXX::INS log_sK = log(S_invK);
-
-		VecXX::INS d1 = invSigmaRootT * (log_sK + halfSigmaSqrd_t);
-		VecXX::INS d2 = d1 - sigmaRootT;
-		VecXX::INS normD1 = cdfnorm(d1);
-		VecXX::INS normD2 = cdfnorm(d2);
-		VecXX::INS C = S * normD1 - Strike * discountedRate * normD2;
-		return C;
-
-	};
-
-	int j = 2 * N + 1 - 1;
-	int i = 0;
-
-	//use BS transform and normal for first pair of slices
-	even_slice = transform(blackScholeInit, terminalAssetPrices);
-
-	//even_slice = transform(payOffFunc, terminalAssetPrices);
-
-	//std::vector<double> dbg = even_slice;
-
-	transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-
-	i += 2;
-	j -= 2;
-
-	for (; i < N; i += 2)
-	{
-		transform(odd_slice, even_slice, trinomialRollBack, sampler, i, j);
-		// transform to get early excercise for american bit , identity sampler just passes values straight through
-		//transform(even_slice, excerciseValue, even_slice, applyEarlyExcercise, identity_sampler, i, j);
-
-		transform(even_slice, odd_slice, trinomialRollBack, sampler, i + 1, j - 1);
-		//transform(odd_slice, excerciseValue, odd_slice, applyEarlyExcercise, identity_sampler, i + 1, j - 1);
-
-		j -= 2;
-	}
-
-    ignore(applyEarlyExcercise);
-
-	return odd_slice[N];
-
 }
 
 
@@ -1719,62 +289,8 @@ void doTrinomialPricerWithInit()
 }
 
 
-
-void doMatrix()
-{
-	VecXX owningVec(0.1, 16*10);
-
-	auto val = 0.0;
-	for (auto& x : owningVec)
-	{
-		x = val;
-		val++;
-	}
-
-	using MAT1 =  Layout2D<double, 8, 0>;
-
-	//auto pDat =
-	MDSpan<double,MAT1> mat(owningVec.data(), 10, 10);
-
-	for (int i = 0; i < 10; ++i)
-	{
-		for (int j = 0; j < 10; ++j)
-		{
-			std::cout << mat(i, j) << ",";
-		}
-		std::cout << "\n";
-	}
-
-	for (int i = 0; i < 10; ++i)
-	{
-		for (int j = 0; j < 10; ++j)
-		{
-			mat(i, j) = i * 100 + j;
-		}
-		
-	}
-
-	std::cout << "\n";
-	std::cout << "\n";
-
-	for (int i = 0; i < 10; ++i)
-	{
-		for (int j = 0; j < 10; ++j)
-		{
-			std::cout << mat(i, j) << ",";
-		}
-		std::cout << "\n";
-	}
-
-}
-
-
-
 int main()
 {
-	//doTransformWithASpan();
-
-	//doStridedSpan();
 	//return 0;
 
 	doBinomialPricer();
@@ -1792,18 +308,16 @@ int main()
 	doAmericanCrankNicholson();
 	//return 0;
 
+	doAmericanTrinomialPricerUpAndOut();
 
-   //doAmericanTrinomialPricerUpAndOut();
-	//return 0;
-
+	//lattice bits
 	doZipping();
-	//	return 0;
 
 	doScan();
-	
+		
 	doTransformWithASpan();
-//	return 0;
 
+	doStridedSpan();
 
 	doMatrix();
 	return 0;
