+++ lattice/lattice_tools.h
@@ -1,0 +1,15 @@
+#pragma once
+
+//lattice tools
+
+void doScan();
+
+void testSampler();
+
+void doStridedSpan();
+
+void doTransformWithASpan();
+
+void doZipping();
+
+void doMatrix();
